Re-add lwip as git submodule
Checked out to lastest release as it includes fix applied in Matter
repository as well as some stuff referred to in repository issues.

PBUF_POOL_BUFSIZE does not use sizeof anymore (it is used in one if in
the lwip upstream code). Additionally removed turning off sanity checks
that were necessary due to aforementioned sizeof usage.

Added LwIPLog to Bouffalolab Logging as I couldnt find it defined
anywhere for this platform.

Signed-off-by: Alexander Mazuruk <aalexanderr@gmail.com>